pyproject.toml
[project]
name = "ai-knowledge-graph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "openai>=1.106.1",
    "python-dotenv>=1.1.1",
    "pyvis>=0.3.2",
    "requests>=2.32.5",
]
